Voyage-Trip-Planner/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
tavily-python>=0.3.0
pydantic>=2.5.0
streamlit>=1.39.0
plotly>=5.18.0
requests>=2.31.0
python-dotenv>=1.0.0
tenacity>=8.2.3
diskcache>=5.6.3
streamlit-extras>=0.4.0
altair>=5.2.0
pyowm>=3.3.0